pyproject.toml
[project]
name = "1rg-bot"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.61",
    "discord-py>=2.5.2",
]
